Major restructurings for easier extension.
