+++ heat2arm/translators/base.py
@@ -18,6 +18,14 @@
 """
 
 
+def get_ref_heat_resource(heat_resource, property_name):
+    """ get_ref_heat_resource is a helper function which returns the property
+    with the given name from the given Heat resource object.
+    """
+    resource_name = heat_resource.properties.data[property_name].args
+    return heat_resource.stack[resource_name]
+
+
 class BaseHeatARMTranslator(object):
     """ BaseHeatARMTranslator is the base class for all heat to ARM translators
 
@@ -32,6 +40,13 @@ class BaseHeatARMTranslator(object):
     def __init__(self, heat_resource):
         self._heat_resource = heat_resource
         self._name = self._heat_resource.name
+
+    def _make_var_name(self, var):
+        """ _get_var_name is a helper method which constructs the
+        name to be used as an ARM template variable.
+        The result is of the form <parameter_name>_<self._name>.
+        """
+        return "%s_%s" % (var, self._name)
 
     def get_parameters(self):
         """ get_parameters returns the dict of ARM template parameters
@@ -57,11 +72,3 @@ class BaseHeatARMTranslator(object):
         into the resulting ARM template format.
         """
         return []
-
-
-def get_ref_heat_resource(heat_resource, property_name):
-    """ get_ref_heat_resource is a helper function which returns the property
-    with the given name from the given Heat resource object.
-    """
-    resource_name = heat_resource.properties.data[property_name].args
-    return heat_resource.stack[resource_name]
