+++ heat2arm/translators/instances/__init__.py
@@ -1,0 +1,4 @@
+"""
+    This module contains definitions for both the Nova and
+    Cloud Formation instance translators.
+"""
