+++ heat2arm/translators/networking/__init__.py
@@ -1,0 +1,4 @@
+"""
+    This module contains the definition for various Neutron and
+    Cloud Formation resource translators.
+"""
